refactor: Ranking Service 분리
